requirements.txt
streamlit==1.38.0
tensorflow==2.13.1
numpy==1.24.4
pillow==10.4.0
pandas==2.0.3
h5py==3.10.0
